+++ src/main/java/com/oop_paint/Shapes/Square.java
@@ -4,11 +4,11 @@ import com.fasterxml.jackson.annotation.JsonTypeName;
 
 @JsonTypeName("square")
 public class Square extends Shape {
-    private int width;
-    private int height;
-    private int []attributes = new int [2];
+    private double width;
+    private double height;
+    private double []attributes = new double [2];
 
-    public Square(int x, int y, String color, String id, int width, int height) {
+    public Square(double x, double y, String color, String id, double width, double height) {
         super.setX(x);
         super.setY(y);
         super.setFill(color);
@@ -20,19 +20,19 @@ public class Square extends Shape {
     public Square() {
     }
 
-    public int getWidth() {
+    public double getWidth() {
         return width;
     }
 
-    public void setWidth(int width) {
+    public void setWidth(double width) {
         this.width = width;
     }
 
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
 
-    public void setHeight(int height) {
+    public void setHeight(double height) {
         this.height = height;
     }
 
@@ -49,7 +49,7 @@ public class Square extends Shape {
         return attributes;
     }
 
-    public void move(int x, int y){
+    public void move(double x, double y){
         super.setX(x);
         super.setY(y);
     }
